requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2
django-appconf==1.0.4
django-imagekit==4.0.2
idna==2.10
pilkit==2.0
Pillow==8.1.0
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
